Fixed precision of calculations, added new tests for SegmentationAlgorithm

+++ plugin/src/mosaic/filamentSegmentation/GaussPsf.java
@@ -9,18 +9,18 @@ public class GaussPsf {
 		double middley = (double)(yl-1)/2;
 		
 		// Generate values
-		float sum = 0;
-			for (int y = 0; y < yl; ++y) 
-				for (int x = 0; x < xl; ++x) {
-				double val = Math.exp(-(Math.pow(x-middlex, 2) + Math.pow(y-middley, 2))/(2 * sigma * sigma));
-				psf[y][x] = val;
-				sum += val;
-			}
+		double sum = 0;
+		for (int y = 0; y < yl; ++y) 
+			for (int x = 0; x < xl; ++x) {
+			double val = Math.exp(-(Math.pow(x-middlex, 2) + Math.pow(y-middley, 2))/(2 * sigma * sigma));
+			psf[y][x] = val;
+			sum += val;
+		}
 		
 		// Normalize
-			for (int y = 0; y < yl; ++y) {
-				for (int x = 0; x < xl; ++x) {
-				psf[y][x] /= sum;
+		for (int y = 0; y < yl; ++y) {
+			for (int x = 0; x < xl; ++x) {
+			    psf[y][x] /= sum;
 			}
 		}
